Print full error strings on a single line for PyCharm integration
This little change in output allows for the configuration of PyCharm File
Watchers, that can pick up the error messages corresponding to specific
filenames and lines through a regular expression and to display them when
hovering over the error stripe marks in the editor.

The exact configuration details are laid out in the README.md.